+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: carmegon <carmegon@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/24 19:14:34 by carmegon          #+#    #+#             */
-/*   Updated: 2025/05/26 18:06:04 by carmegon         ###   ########.fr       */
+/*   Updated: 2025/05/28 10:28:16 by carmegon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_print_str(char *str);
 int		ft_print_char(int c);
 int		ft_print_nbr(int num);
 int		ft_print_hex_top(unsigned int num);
-int		ft_print_hex_bottom(unsigned int num);
+int		ft_print_hex_bottom(unsigned long long num);
 int		ft_print_ptr(void *ptr);
 int		ft_print_uns(unsigned int num);
